+++ ijkmedia/ijkplayer/ff_ffplay_options.h
@@ -212,7 +212,8 @@ static const AVOption ffp_context_options[] = {
         OPTION_OFFSET(ijkmeta_delay_init),      OPTION_INT(0, 0, 1) },
     { "render-wait-start",          "render wait start",
         OPTION_OFFSET(render_wait_start),      OPTION_INT(0, 0, 1) },
-
+    { "cover-after-prepared",          "display the first video frame as cover, then toggle pause",
+      OPTION_OFFSET(cover_after_prepared),      OPTION_INT(0, 0, 1) },
     { NULL }
 };