+++ ios/IJKMediaPlayer/IJKMediaPlayer/ijkmedia/ijkplayer/ios/ijkplayer_ios.h
@@ -21,7 +21,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
  */
 
-#include "ijkplayer/ijkplayer.h"
+#include "ijkplayer.h"
 #import "IJKSDLGLView.h"
 
 #if IJK_IOS